+++ lib/shared/src/index.ts
@@ -63,6 +63,7 @@ export type {
     CodyCommand,
     CodyCommandContext,
     CodyCommandType,
+    CodyCommandMode,
     TerminalOutputArguments,
 } from './commands/types'
 export { CustomCommandType } from './commands/types'
@@ -127,6 +128,7 @@ export {
     CompletionStopReason,
     type CodeCompletionsClient,
     type CodeCompletionsParams,
+    type SerializedCodeCompletionsParams,
     type CompletionResponseGenerator,
 } from './inferenceClient/misc'
 export type {
@@ -193,7 +195,7 @@ export {
     type BrowserOrNodeResponse,
     type GraphQLAPIClientConfig,
     type LogEventMode,
-    type ContextFiltersResult,
+    type ContextFilters,
     type CodyContextFilterItem,
     type RepoListResponse,
 } from './sourcegraph-api/graphql/client'
@@ -235,10 +237,9 @@ export {
     scanForMentionTriggerInUserTextInput,
 } from './mentions/query'
 export {
-    getURLContextItems,
-    isURLContextItem,
-    fetchContentForURLContextItem,
-} from './mentions/urlContextItems'
+    CONTEXT_MENTION_PROVIDERS,
+    type ContextMentionProvider,
+} from './mentions/api'
 export { TokenCounter } from './token/counter'
 export {
     EXPERIMENTAL_USER_CONTEXT_TOKEN_BUDGET,
@@ -250,3 +251,4 @@ export { getCompletionsModelConfig } from './llm-providers/utils'
 export type { SourcegraphNodeCompletionsClient } from './sourcegraph-api/completions/nodeClient'
 export * from './fetch'
 export * from './completions/types'
+export * from './cody-ignore/context-filters-provider'