+++ vscode/src/commands/menus/command-builder.ts
@@ -2,11 +2,11 @@ import { type QuickPickItem, window } from 'vscode'
 
 import type { CodyCommand } from '@sourcegraph/cody-shared'
 
-import { CustomCommandType } from '@sourcegraph/cody-shared'
+import { type CodyCommandMode, CustomCommandType } from '@sourcegraph/cody-shared'
 import { telemetryService } from '../../services/telemetry'
 import { telemetryRecorder } from '../../services/telemetry-v2'
 import { fromSlashCommand } from '../utils/common'
-import { customPromptsContextOptions } from './items'
+import { CommandModeMenuOptions, customPromptsContextOptions } from './items/menu'
 
 export interface CustomCommandsBuilder {
     key: string
@@ -15,33 +15,45 @@ export interface CustomCommandsBuilder {
 }
 
 export class CustomCommandsBuilderMenu {
+    /**
+     * Starts the process of creating a new custom Cody command.
+     *
+     * This method prompts the user to enter a command name, select a command mode, and enter a prompt for Cody to follow.
+     * If all the required information is provided, it returns a `CustomCommandsBuilder` object that can be used to create the new custom command.
+     *
+     * @param commands - An array of existing command names to check for duplicates.
+     * @returns A `CustomCommandsBuilder` object if the command creation process is successful, or `null` if the user cancels or the input is invalid.
+     */
     public async start(commands: string[]): Promise<CustomCommandsBuilder | null> {
         const key = await this.makeCommandKey(commands)
-        if (!key) {
-            return null
+        const mode = key && (await this.selectCommandMode())
+        const prompt = mode && (await this.getCommandPrompt())
+        const type = prompt && (await this.selectCommandType())
+
+        if (key && mode && prompt && type) {
+            telemetryService.log('CodyVSCodeExtension:command:custom:build:executed')
+            telemetryRecorder.recordEvent('cody.command.custom.build', 'executed')
+
+            return { key, prompt: { ...prompt, key, mode }, type }
         }
 
-        const prompt = await this.makePrompt()
-        if (!prompt) {
-            return null
-        }
-
-        const type = await this.makeType()
-        if (!type) {
-            return null
-        }
-
-        telemetryService.log('CodyVSCodeExtension:command:custom:build:executed')
-        telemetryRecorder.recordEvent('cody.command.custom.build', 'executed')
-        return { key, prompt: { ...prompt, key }, type }
+        return null
     }
 
+    /**
+     * STEP 1: Make a new command key
+     * Prompts the user to enter a name for a new custom Cody command, validating that the name is not empty and does not contain spaces.
+     * It also checks that the command name does not already exist in the list of commands.
+     *
+     * @param commands - An array of existing command names.
+     * @returns The new command name entered by the user, or `undefined` if the user cancels the input.
+     */
     private async makeCommandKey(commands: string[]): Promise<string | undefined> {
         const commandSet = new Set(commands)
         const value = await window.showInputBox({
             title: 'New Custom Cody Command: Command Name',
-            prompt: 'Enter the name of the custom command.',
-            placeHolder: 'e.g. hello',
+            prompt: 'Enter a unique keyword for the command.',
+            placeHolder: 'e.g. spellchecker',
             ignoreFocusOut: true,
             validateInput: (input: string) => {
                 if (!input) {
@@ -61,31 +73,60 @@ export class CustomCommandsBuilderMenu {
         return value
     }
 
-    private async makePrompt(): Promise<Omit<CodyCommand, 'key'> | null> {
+    /**
+     * STEP 2: Select a command mode
+     * Displays a quick pick menu to allow the user to select a command mode for a new custom Cody command.
+     * The command mode determines how the command will be executed, such as whether it should run in the current context or in a separate process.
+     *
+     * @returns The selected command mode, or `undefined` if the user cancels the selection.
+     */
+    private async selectCommandMode(): Promise<CodyCommandMode | undefined> {
+        const commandMode = await window.showQuickPick(CommandModeMenuOptions, {
+            title: 'New Custom Cody Command: Command Mode',
+            placeHolder: 'Choose how the command should be executed...',
+            canPickMany: false,
+            onDidSelectItem: (item: QuickPickItem) => {
+                item.picked = !item.picked
+            },
+        })
+
+        return commandMode ? (commandMode.id as CodyCommandMode) : undefined
+    }
+
+    /**
+     * STEP 3: Enter a new prompt for the command
+     * Prompts the user to enter a new custom Cody command prompt and adds the necessary context options for the command.
+     *
+     * @returns The new custom Cody command with the prompt and context, or `null` if the user cancels the operation.
+     */
+    private async getCommandPrompt(): Promise<Omit<CodyCommand, 'key'> | null> {
         const prompt = await window.showInputBox({
             title: 'New Custom Cody Command: Prompt',
             prompt: 'Enter the instructions for Cody to follow and answer.',
             placeHolder: 'e.g. Create five different test cases for the selected code',
-            ignoreFocusOut: true,
-            validateInput: (input: string) => {
-                if (!input) {
-                    return 'Command prompt cannot be empty.'
-                }
-                return null
-            },
+            ignoreFocusOut: false,
+            validateInput: (input: string) => (input.length ? null : 'Command prompt cannot be empty.'),
         })
-        if (!prompt) {
-            return null
-        }
-        return this.addContext({ prompt })
+        return prompt ? this.addContext({ prompt }) : null
     }
 
-    private async addContext(newPrompt?: Partial<CodyCommand>): Promise<CodyCommand | null> {
-        if (!newPrompt) {
-            return null
-        }
-
-        newPrompt.context = { ...{ codebase: false } }
+    /**
+     * STEP 4: Add context to the new Cody command
+     * Adds context options to the new Cody command.
+     * This function allows the user to select which context options to include in the new Cody command. The available context options are:
+     * - Selection: Include the current text selection in the prompt context.
+     * - Current Directory: Include the current working directory in the prompt context.
+     * - Open Tabs: Include the currently open tabs in the prompt context.
+     * - None: Include no additional context in the prompt.
+     * - Command: Allow the user to enter a terminal command to run from the workspace root, and include its output in the prompt context.
+     *
+     * If the user does not select any context options, the function will return the new Cody command without any additional context.
+     *
+     * @param newPrompt - The new Cody command to add context to.
+     * @returns The new Cody command with the selected context options, or null if no prompt was provided.
+     */
+    private async addContext(newPrompt: Partial<CodyCommand>): Promise<CodyCommand | null> {
+        newPrompt.context = {}
         const promptContext = await window.showQuickPick(customPromptsContextOptions, {
             title: 'New Custom Cody Command: Context Options',
             placeHolder: 'For accurate responses, choose only the necessary options.',
@@ -96,21 +137,19 @@ export class CustomCommandsBuilderMenu {
             },
         })
 
-        if (!promptContext?.length) {
-            return newPrompt as CodyCommand
-        }
-
-        for (const context of promptContext) {
-            switch (context.id) {
-                case 'selection':
-                case 'currentDir':
-                case 'openTabs':
-                case 'none':
-                    newPrompt.context[context.id] = context.picked
-                    break
-                case 'command': {
-                    newPrompt.context.command = (await showPromptCreationInputBox()) || ''
-                    break
+        if (promptContext !== undefined) {
+            for (const context of promptContext) {
+                switch (context.id) {
+                    case 'selection':
+                    case 'currentDir':
+                    case 'openTabs':
+                    case 'none':
+                        newPrompt.context[context.id] = context.picked
+                        break
+                    case 'command': {
+                        newPrompt.context.command = (await showPromptCreationInputBox()) ?? undefined
+                        break
+                    }
                 }
             }
         }
@@ -118,7 +157,12 @@ export class CustomCommandsBuilderMenu {
         return newPrompt as CodyCommand
     }
 
-    private async makeType(): Promise<CustomCommandType> {
+    /**
+     * STEP 5: Save the new Cody command
+     * Prompts the user to choose whether to save a custom Cody command to the user or workspace settings.
+     * @returns A promise that resolves to the chosen `CustomCommandType`.
+     */
+    private async selectCommandType(): Promise<CustomCommandType> {
         const option = await window.showQuickPick(
             [
                 {
@@ -141,6 +185,10 @@ export class CustomCommandsBuilderMenu {
                 placeHolder: 'Choose where to save the command',
             }
         )
+
+        if (!option?.type) {
+            throw new Error('Custom Command creation aborted.')
+        }
 
         return option?.type === CustomCommandType.Workspace
             ? CustomCommandType.Workspace