+++ vscode/test/e2e/command-custom.test.ts
@@ -11,6 +11,7 @@ import {
 import {
     type DotcomUrlOverride,
     type ExpectedEvents,
+    openCustomCommandMenu,
     test as baseTest,
     withPlatformSlashes,
 } from './helpers'
@@ -55,7 +56,7 @@ test.extend<ExpectedEvents>({
     // Bring the cody sidebar to the foreground
     await page.click('.badge[aria-label="Cody"]')
     // Click the Custom Commands button in the Sidebar to open the Custom Commands menu
-    await page.getByText('Custom Commands').click()
+    await page.getByText('Custom Commands', { exact: true }).click()
 
     const commandName = 'ATestCommand'
     const prompt = 'The test command has been created'
@@ -66,23 +67,33 @@ test.extend<ExpectedEvents>({
         .locator('a')
         .filter({ hasText: /New Custom Command.../ })
         .click()
+
     // Enter command name
     const commandInputTitle = page.getByText('New Custom Cody Command: Command Name')
     await expect(commandInputTitle).toBeVisible()
-    const commandInputBox = page.getByPlaceholder('e.g. hello')
+    const commandInputBox = page.getByPlaceholder('e.g. spellchecker')
     await commandInputBox.fill(commandName)
     await commandInputBox.press('Enter')
+
+    // Select mode
+    const commandModeTitle = page.getByText('New Custom Cody Command: Command Mode')
+    await expect(commandModeTitle).toBeVisible()
+    // Hit enter to select the first option on the list: 'ask'
+    await page.keyboard.press('Enter')
+
     // Enter prompt
     const promptInputTitle = page.getByText('New Custom Cody Command: Prompt')
     await expect(promptInputTitle).toBeVisible()
     const promptInputBox = page.getByPlaceholder(
         'e.g. Create five different test cases for the selected code'
     )
+
     await promptInputBox.fill(prompt)
     await promptInputBox.press('Enter')
     // Use default context
     await expect(page.getByText('New Custom Cody Command: Context Options')).toBeVisible()
     await page.keyboard.press('Enter')
+
     // Save it to workspace settings
     await expect(page.getByText('New Custom Cody Command: Save To…')).toBeVisible()
     await expect(page.getByText('Workspace Settings.vscode/cody.json')).toBeVisible()
@@ -112,7 +123,7 @@ test.extend<ExpectedEvents>({
 
     // Show the new command in the menu and execute it
     await page.click('.badge[aria-label="Cody"]')
-    await page.getByLabel('Custom Commands').locator('a').click()
+    await openCustomCommandMenu(page)
     await page.getByText('Cody: Custom Commands (Beta)').hover()
     await expect(page.getByText('Cody: Custom Commands (Beta)')).toBeVisible()
     await page.getByPlaceholder('Search command to run...').click()
@@ -152,14 +163,12 @@ test.extend<ExpectedEvents>({
     // Wait for index.html to fully open
     await page.getByRole('tab', { name: 'index.html' }).hover()
 
-    // Bring the cody sidebar to the foreground
-    await page.click('.badge[aria-label="Cody"]')
-
     // Open the chat sidebar to click on the Custom Command option
     // Search for the command defined in cody.json and execute it
+    await page.click('.badge[aria-label="Cody"]')
+    await openCustomCommandMenu(page)
 
     /* Test: context.currentDir with currentDir command */
-    await page.getByRole('treeitem', { name: 'Custom Commands' }).locator('a').click()
     await expect(page.getByPlaceholder('Search command to run...')).toBeVisible()
     await page.getByPlaceholder('Search command to run...').fill('currentDir')
     await page.keyboard.press('Enter')
@@ -189,7 +198,7 @@ test.extend<ExpectedEvents>({
 
     /* Test: context.directory with directory command */
 
-    await page.getByRole('treeitem', { name: 'Custom Commands' }).locator('a').click()
+    await openCustomCommandMenu(page)
     await expect(page.getByPlaceholder('Search command to run...')).toBeVisible()
     await page.getByPlaceholder('Search command to run...').click()
     await page.getByPlaceholder('Search command to run...').fill('directory')
@@ -209,7 +218,7 @@ test.extend<ExpectedEvents>({
 
     /* Test: context.openTabs with openTabs command */
 
-    await page.getByRole('treeitem', { name: 'Custom Commands' }).locator('a').click()
+    await openCustomCommandMenu(page)
     await expect(page.getByPlaceholder('Search command to run...')).toBeVisible()
     await page.getByPlaceholder('Search command to run...').click()
     await page.getByPlaceholder('Search command to run...').fill('openTabs')
@@ -253,14 +262,9 @@ test.extend<ExpectedEvents>({
     await page.getByRole('tab', { name: 'cody.json' }).hover()
 
     await page.click('.badge[aria-label="Cody"]')
-
-    // Check button click to open the cody.json file in the editor
-    // const label = 'gear  Configure Custom Commands..., Manage your custom reusable commands, settings'
-    // const configMenuItem = page.getByLabel(label).locator('a')
-    const customCommandSidebar = page.getByRole('treeitem', { name: 'Custom Commands' }).locator('a')
+    await openCustomCommandMenu(page)
 
     // Able to open the cody.json file in the editor from the command menu
-    await customCommandSidebar.click()
     await expect(page.getByPlaceholder('Search command to run...')).toBeVisible()
     await page.getByLabel('Configure Custom Commands...', { exact: true }).click()
     await page.locator('a').filter({ hasText: 'Open Workspace Settings (JSON)' }).hover()
@@ -274,7 +278,7 @@ test.extend<ExpectedEvents>({
     await codyJSONFileTab.getByRole('button', { name: /^Close/ }).click()
 
     // Check button click to delete the cody.json file from the workspace tree view
-    await customCommandSidebar.click()
+    await openCustomCommandMenu(page)
     await expect(page.getByPlaceholder('Search command to run...')).toBeVisible()
     await page.getByLabel('Configure Custom Commands...', { exact: true }).click()
     await page.locator('a').filter({ hasText: 'Open Workspace Settings (JSON)' }).hover()
@@ -294,7 +298,7 @@ test.extend<ExpectedEvents>({
     // NOTE: This is expected to fail locally if you currently have User commands configured
     await page.waitForTimeout(100)
     await page.click('.badge[aria-label="Cody"]')
-    await customCommandSidebar.click()
+    await openCustomCommandMenu(page)
     await page.locator('a').filter({ hasText: 'Open User Settings (JSON)' }).hover()
     await page.getByRole('button', { name: 'Open or Create Settings File' }).hover()
     await page.getByRole('button', { name: 'Open or Create Settings File' }).click()
@@ -318,6 +322,8 @@ testGitWorkspace('use terminal output as context', async ({ page, sidebar }) => 
     await expect(menuInputBox).toBeVisible()
     await menuInputBox.fill('shellOutput')
     await page.keyboard.press('Enter')
+
+    await expect(menuInputBox).not.toBeVisible()
 
     // Check the context list to confirm the terminal output is added as file
     const panel = getChatPanel(page)